Refactor unit tests by using for comprehension
Unit tests have been refactored by utilising for comprehensions and importing ExecutionContext.Implicits.global. Purpose of this change is to simplify the code. Consequentially, avoiding nesting `whenReady` which felt cumbersome and dirty. By using flatMap via for comprehension, the structure is more clean and ordered. In the process, also increased the `patienceConfig` to 1 second for less brittle behavior.