Remove unnecessary logging from ChanNullDAOTest
This commit removes the import and usage of the Logging class in the ChanNullDAOTest file, as the logging statement was being used only once to display a page of Channel Nulls in tests. To keep our tests clean and focused on testing functionality instead of log outputs, this logging statement has been removed.