Implement user blocking feature in posts retrieval
The methods in ‘ChanNullPostDAO.scala’ and ‘ChanNullPostDAOImpl.scala’ were modified to add a ‘loggedInUserId’ parameter. This allows the system to check for any user-specific blocks when retrieving posts. The implemented blocking feature will filter posts created by users blocked by the 'loggedInUserId'. This change affects the way posts are retrieved, as it brings user-specific context into the retrieval process.